backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
python-dotenv
xgboost
pandas
scikit-learn
joblib
google-genai
openai
groq
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-docx
python-multipart
numpy
httpx
beautifulsoup4
lxml
